84: Splitting app into components
