+++ src/components/Cockpit/Cockpit.module.css
@@ -1,0 +1,20 @@
+.red {
+    color: red;
+}
+
+.bold {
+    font-weight: bold;
+}
+
+.switchPersonsButton {
+    background-color: white;
+    font: inherit;
+    border: 1px solid blue;
+    padding: 8px;
+    cursor: pointer;
+}
+
+.switchPersonsButton:hover {
+    background-color: aquamarine;
+    color: blue;
+}
