+++ src/containers/App.module.css
@@ -2,23 +2,3 @@
     text-align: center;
   }
   
-.red {
-    color: red;
-}
-
-.bold {
-    font-weight: bold;
-}
-
-.switchPersonsButton {
-    background-color: white;
-    font: inherit;
-    border: 1px solid blue;
-    padding: 8px;
-    cursor: pointer;
-}
-
-.switchPersonsButton:hover {
-    background-color: aquamarine;
-    color: blue;
-}
